plumb kube.get and introduce new KindFixture
Reworks the processing of the kubeconfig, kubecontext, and namespace to
be more reasonable and readable as well as to take into account any
fixtures which expose `kube.config` or `kube.context` state keys.

Adds a new `fixtures/kind:KindFixture` that can either notify gdt about
an existing KinD cluster or create a new one.

Finally, implements the `kube.get` field with basic assertions about
List or Get response size, the ability to check for an error in the
response and a shortcut assertion for checking lack of a resource's
existence.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>