tighten up resource discovery/client calls
Moves the client/config `Spec` methods into a new `connect.go` file and
introduces a helper `Spec.connect()` method that returns a `connection`
struct that houses the discovery and dynamic client, along with a cache
of preferred API versions for API groups.

It took a long time and a bunch of trial-and-error to figure out how to
work with the apimachinery package's `apis/meta/v1` types like
`APIResource`, `APIResourceList` and `APIGroup` along with the
apimachinery package's `runtime/schema` types like `GroupVersionKind and
`GroupVersionResource`. Why on God's green Earth these basic types need
to be so convoluted and duplicative is beyond me.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>