adds new `Kube.assert.matches` functionality
Adds the ability to specify a match resource or set of fields to compare
against the result from a `kube.get` call.

Matches is either a string or a map[string]interface{} containing the
resource that the `Kube.Get` should match against. If Matches is a
string, the string can be either a file path to a YAML manifest or
inline an YAML string containing the resource fields to compare.

Only fields present in the Matches resource are compared. There is a
check for existence in the retrieved resource as well as a check that
the value of the fields match. Only scalar fields are matched entirely.
In other words, you do not need to specify every field of a struct field
in order to compare the value of a single field in the nested struct.

As an example, imagine you wanted to check that a Deployment resource's
`Status.ReadyReplicas` field was 2. You do not need to specify all other
`Deployment.Status` fields like `Status.Replicas` in order to match the
`Status.ReadyReplicas` field value. You only need to include the
`Status.ReadyReplicas` field in the `Matches` value as these examples
demonstrate:

```yaml
tests:
 - name: check deployment's ready replicas is 2
   kube:
     get: deployments/my-deployment
     assert:
       matches: |
         kind: Deployment
         metadata:
           name: my-deployment
         status:
           readyReplicas: 2
```

you don't even need to include the kind and metadata in `Matches`. If
missing, no kind and name matching will be performed.

```yaml
tests:
 - name: check deployment's ready replicas is 2
   kube:
     get: deployments/my-deployment
     assert:
       matches: |
         status:
           readyReplicas: 2
```

In fact, you don't need to use an inline multiline YAML string. You can
use a `map[string]interface{}` as well:

```yaml
tests:
 - name: check deployment's ready replicas is 2
   kube:
     get: deployments/my-deployment
     assert:
       matches:
         status:
           readyReplicas: 2
```

Signed-off-by: Jay Pipes <jaypipes@gmail.com>