add support for `kube.assert.json` assertions
Adds support for specifying expected field, field value, and field value
formats using the `gdt-core` `assertion/json.Expect` struct.

Expectations of field to find (and an expected value of that field) in a
Kubernetes resource can now be specified using `kube.assert.json.paths`,
like this example illustrates:

```yaml
tests:
 - name: check deployment's ready replicas is 2
   kube:
     get: deployments/my-deployment
     assert:
       json:
        paths:
          $.status.readyReplicas: 2
```

JSONPath expressions can be fairly complex, allowing the test author to, for
example, assert the value of a nested map field with a particular key, as this
example shows:

```yaml
tests:
 - name: check deployment's pod template "app" label is "nginx"
   kube:
     get: deployments/my-deployment
     assert:
       json:
        paths:
          $.spec.template.labels["app"]: nginx
```

You can check that the value of a particular field at a JSONPath is formatted
in a particular fashion using `kube.assert.json.path_formats`. This is a map,
keyed by JSONPath expression, of the data format the value of the field at that
JSONPath expression should have.

For example, suppose we wanted to verify that a Deployment's `metadata.uid`
field was a UUID-4 and that its `metadata.creationTimestamp` field was a
date-time timestamp:

```yaml
tests:
  - kube:
      get: deployments/nginx
      assert:
        json:
          path_formats:
            $.metadata.uid: uuid4
            $.metadata.creationTimestamp: date-time
```

Closes Issue #5

Signed-off-by: Jay Pipes <jaypipes@gmail.com>